% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dl_data_from}
\alias{dl_data_from}
\title{Downloads the news about cryptocurrencies from CryptoCompare API}
\usage{
dl_data_from(timestp)
}
\arguments{
\item{timestamp}{a character of numerical value representing time}
}
\value{
Dataframe with the time and body of the news released after timestamp
}
\description{
Downloads the news about cryptocurrencies from CryptoCompare API
}
\examples{
dl_data_from(1543820423)
}
