% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Tsp_data}
\alias{Tsp_data}
\title{Downloads the news about cryptocurrencies from CryptoCompare API}
\usage{
Tsp_data(timestamp)
}
\arguments{
\item{timestamp}{a character of numerical value representing time}
}
\value{
Dataframe with the time and body of the 50 latests news released before timestamp
}
\description{
Downloads the news about cryptocurrencies from CryptoCompare API
}
\examples{
Tsp_data(1543920423)
}
