% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{updateDayNewsData}
\alias{updateDayNewsData}
\title{Updates the dataset CryptoNewsAnalysedDays with the latest news from CryptoCompare API}
\usage{
updateDayNewsData()
}
\description{
Updates the dataset CryptoNewsAnalysedDays with the latest news from CryptoCompare API
}
\examples{
updateDayNewsData()
}
