% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{updateHourNewsData}
\alias{updateHourNewsData}
\title{Updates the dataset CryptoNewsAnalysedHours with the latest news from CryptoCompare API}
\usage{
updateHourNewsData()
}
\description{
Updates the dataset CryptoNewsAnalysedHours with the latest news from CryptoCompare API
}
\examples{
updateHourNewsData()
}
